Add files via upload
Added User Stories
(As a user, I want to hide the application GUI, so that it does not block my view when I am not using it. 
As a user, I want to reopen the application GUI, so that I can access the settings when I need to(when webcam starts hide the main application, then maybe press a button it will reappear?).